requirements.txt
python-decouple
requests
yaml